fix: round-trip Volume dB edits and anchor mute at the floor (#42)
The Players-tab dB experience had three defects surfaced on a re-provisioned
streamer:

- dB slider re-seeded at a different spot after an edit, because
  db_to_percent truncated to int and DSPConfig.volume was an int, so a dB
  edit round-tripped lossily.
- A normal mid-range dB edit silently muted the client: db_to_percent
  truncated anything below ~-40.5 dB to 0, and the persist path muted
  whenever percent == 0.

Fixes:
- db_to_percent returns a precise float (0.0 only at/below the floor);
  DSPConfig.volume is now a float so dB edits round-trip without drift.
- MIN_DB -80.0 -> -50.0 to match CamillaGUI's default slider range.
- Mute is anchored at the slider floor (db <= MIN_DB / percent <= 0), not
  on lossy zero-rounding.
- Both slider seeds are step-aligned so the periodic refresh does not fire
  a phantom update:model-value.

+++ audera/models/dsp.py
@@ -34,7 +34,7 @@ class DSPConfig(BaseModel):
     enabled: bool = True
     loudness_enabled: bool = False
     loudness_reference_level: float = -25.0
-    volume: int = 25
+    volume: float = 25.0
 
     @field_validator('loudness_reference_level', mode='before')
     @classmethod
